requirements.txt
pyTelegramBotAPI>=4.12.0
pymongo[srv]>=4.5.0
python-dotenv>=1.0.0
certifi>=2023.7.22